Move terminal package
fixes issue where terminal_check_notappengine.go can't access terminal
package since terminal package is in an internal package
