+++ terminal/terminal_check_bsd.go
