+++ terminal/terminal_check_unix.go
